link to summary of Imperial College report
The report is produced by one of the most respected rapid pandemic response units in the world.  Models the impact and spared of COVID-19 under different intervention strategies.  It's grim, but I thought it would be a good reference in this section. Thought to have directly influence the UK and US change of course in the response (certainly the UK side). If anything, the messages in the report are not being shared widely enough.